Estoy arreglando detalles chiquitos que rompen, al pararse y definir el ganador ya le suma la apuesta al saldo si el jugador gano, en el caso de que este dividida va sumando la mitad en cada mano (por si una gano y otra no), tambien bloquee el boton de empezar partida y se desbloquea cuando hay por lo menos una apuesta, el boton de dividir tambien duplica la apuesta

+++ src/main/java/com/tallerwebi/dominio/Usuario.java
@@ -14,7 +14,7 @@ public class Usuario {
     @Column(unique = true)
     private String email;
     private String password;
-    private Double saldo = 150.0;
+    private Integer saldo = 150;
     private String nombre;
     private String apellido;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
@@ -38,8 +38,8 @@ public class Usuario {
     public Boolean getAceptoTerminos() {return aceptoTerminos;}
     public String getRol() {return rol;}
     public void setRol(String rol) {this.rol = rol;}
-    public Double getSaldo() { return saldo; }
-    public void setSaldo(Double saldo) { this.saldo = saldo; }
+    public Integer getSaldo() { return saldo; }
+    public void setSaldo(Integer saldo) { this.saldo = saldo; }
     public Long getId() {
         return id;
     }
